refactor: add nested routes to /profile
