+++ src/components/index.js
@@ -14,5 +14,7 @@ export { default as IngredientDetails } from './ingredient-details'
 export { default as CloseButton } from './close-button'
 export { default as ConstructorTile } from './constructor-tile'
 export { default as Loading } from './loading'
+export { default as ProfileForm } from './profile-form'
+export { default as OrderList } from './order-list'
 export { default as ProtectedRoute } from './protected-route'
 export { default as NotFound404 } from './not-found404'
