+++ src/components/profile-form/index.js
@@ -1,0 +1,1 @@
+export { default } from './profile-form'
