+++ src/pages/profile/profile.jsx
@@ -1,65 +1,17 @@
-import { useCallback, useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import { Link, NavLink } from 'react-router-dom'
+import { useCallback } from 'react'
+import { useDispatch } from 'react-redux'
+import { Link, NavLink, Route, Switch, useRouteMatch} from 'react-router-dom'
 
-import {
-  Button,
-  EmailInput,
-  Input,
-  PasswordInput,
-} from '@ya.praktikum/react-developer-burger-ui-components'
-
-import { editUser, signOut } from 'services/actions/auth'
+import { signOut } from 'services/actions/auth'
 
 import profileStyles from './profile.module.scss'
-import { useForm } from 'hooks'
+import ProfileForm from 'components/profile-form/profile-form'
+import { OrderList } from 'components'
 
 const ProfilePage = () => {
-  const { user } = useSelector(store => store.auth)
   const dispatch = useDispatch()
 
-  const { form, setForm, handleChange } = useForm({
-    name: '',
-    email: '',
-    password: '',
-  })
-
-  const resetForm = useCallback(
-    () =>
-      setForm(form => ({
-        name: user?.name || '',
-        email: user?.email || '',
-        password: '',
-      })),
-    [user?.name, user?.email, setForm],
-  )
-
-  const isFormChanged = form.name !== user?.name || form.email !== user?.email
-
-  useEffect(() => {
-    resetForm()
-  }, [resetForm]) //reset form values when stored user changes
-
-  const submitHandler = useCallback(
-    e => {
-      e.preventDefault()
-      dispatch(editUser(form))
-      resetForm()
-    },
-    [dispatch, form, resetForm],
-  )
-
-  const resetHandler = useCallback(
-    e => {
-      e.preventDefault()
-      resetForm()
-    },
-    [resetForm],
-  )
-
-  const handleExit = useCallback(() => {
-    dispatch(signOut())
-  }, [dispatch])
+  const { path } = useRouteMatch()
 
   const generateNavLinkClassname = useCallback(
     isActive =>
@@ -68,6 +20,10 @@ const ProfilePage = () => {
       } pt-4 pb-4`,
     [],
   )
+
+  const handleExit = useCallback(() => {
+    dispatch(signOut())
+  }, [dispatch])
 
   return (
     <div className={`${profileStyles.container} mt-30`}>
@@ -101,57 +57,20 @@ const ProfilePage = () => {
           В этом разделе вы можете изменить свои персональные данные
         </p>
       </div>
-      <form
-        className={profileStyles.form}
-        onSubmit={submitHandler}
-        onReset={resetHandler}
-      >
-        <Input
-          name='name'
-          onChange={handleChange}
-          placeholder='Имя'
-          value={form.name}
-          extraClass='mb-6'
-          icon='EditIcon'
-        />
-        <EmailInput
-          name='email'
-          onChange={handleChange}
-          placeholder='Логин'
-          value={form.email}
-          errorText='Ой! Кажется, в Вашем адресе ошибка :('
-          extraClass='mb-6'
-          icon='EditIcon'
-        />
-        <PasswordInput
-          name='password'
-          onChange={handleChange}
-          placeholder='Пароль'
-          value={form.password}
-          extraClass='mb-6'
-          icon='EditIcon'
-        />
-        {(isFormChanged || form.password) && (
-          <div className={profileStyles.controls}>
-            <Button
-              size='medium'
-              htmlType='submit'
-              type='primary'
-              extraClass='mb-20'
-            >
-              Сохранить
-            </Button>
-            <Button
-              size='medium'
-              htmlType='reset'
-              type='primary'
-              extraClass='mb-20'
-            >
-              Отменить
-            </Button>
-          </div>
-        )}
-      </form>
+      <Switch>
+        <Route
+          path={path}
+          exact
+        >
+          <ProfileForm />
+        </Route>
+        <Route
+          path={`${path}/orders`}
+          exact
+        >
+          <OrderList />
+        </Route>
+      </Switch>
     </div>
   )
 }
